[ refactor ] 불필요한 custom hook 제거
