+++ src/@components/CardCollectionPage/Card/LastCard/index.tsx
@@ -1,10 +1,9 @@
-import React, { forwardRef, useEffect } from "react";
-import { useNavigate } from "react-router-dom";
+import React, { forwardRef, useEffect, useState } from "react";
+import { useNavigate, useSearchParams } from "react-router-dom";
 
 import { IcCongratPiickle } from "../../../../asset/icon";
 import { routePaths } from "../../../../core/routes/path";
 import useToast from "../../../@common/Toast/hooks/useToast";
-import useReplayButton from "../../hooks/useTopicReplay";
 import * as St from "./style";
 
 const TOAST_SESSON_KEY = "showToast";
@@ -12,8 +11,16 @@ const TOAST_SESSON_KEY = "showToast";
 const LastCard = forwardRef(function LastCard(_, ref: React.ForwardedRef<HTMLDivElement>) {
   const navigate = useNavigate();
   const showToast = useToast();
-  const isReplayBtn = useReplayButton();
   const showToastFlag = !!sessionStorage.getItem(TOAST_SESSON_KEY);
+  const [isReplayBtn, setIsReplayBtn] = useState(false);
+  const [searchParams] = useSearchParams();
+  const cardType = searchParams.get("type");
+
+  useEffect(() => {
+    const noRepalyTypes = ["female", "male", "bookmark", "medley", "recent", "best"];
+
+    setIsReplayBtn(!noRepalyTypes.includes(cardType || ""));
+  }, [cardType]);
 
   useEffect(() => {
     if (showToastFlag) {
@@ -22,7 +29,7 @@ const LastCard = forwardRef(function LastCard(_, ref: React.ForwardedRef<HTMLDiv
     }
   }, [showToastFlag, showToast]);
 
-  const getSimilarTopic = () => {
+  const reloadForSimilarTopic = () => {
     sessionStorage.setItem(TOAST_SESSON_KEY, "true");
     window.location.reload();
   };
@@ -35,7 +42,7 @@ const LastCard = forwardRef(function LastCard(_, ref: React.ForwardedRef<HTMLDiv
       <St.Content>끊임없는 대화를 위해 버튼을 선택해주세요</St.Content>
       <IcCongratPiickle />
       <St.BtnContainer>
-        {isReplayBtn && <St.ReplayBtn onClick={getSimilarTopic}>비슷한 주제 계속 보기</St.ReplayBtn>}
+        {isReplayBtn && <St.ReplayBtn onClick={reloadForSimilarTopic}>비슷한 주제 계속 보기</St.ReplayBtn>}
         <St.CategoryBtn onClick={goToCategory}>카테고리 보러 가기</St.CategoryBtn>
       </St.BtnContainer>
     </St.Card>
